refactor(dashboard-links): own shared link read model
